fix: implement UI for desktop mode for demo purpose
