+++ packages/trader/src/AppV2/Components/AccumulatorStats/accumulator-stats.tsx
@@ -71,7 +71,7 @@ const AccumulatorStats = observer(() => {
         return null;
     }
     return (
-        <div>
+        <React.Fragment>
             <div className='accumulators-stats-v2'>
                 <div className='accumulators-stats-v2__container'>
                     <button
@@ -115,7 +115,7 @@ const AccumulatorStats = observer(() => {
                 )}
                 {is_open_description && <AccumulatorStatsDescription onActionSheetClose={onActionSheetClose} />}
             </ActionSheet.Root>
-        </div>
+        </React.Fragment>
     );
 });
