+++ packages/trader/src/AppV2/Containers/ContractDetails/contract-details.scss
@@ -11,7 +11,7 @@ $TWO_BUTTONS_FOOTER_HEIGHT: 120px;
     overflow-y: auto;
     overflow-x: hidden;
     gap: var(--semantic-spacing-general-sm);
-    height: 100%;
+    height: calc(100vh - 40px);
     padding-bottom: var(--semantic-spacing-general-sm);
 
     .contract-details-chart {
