--- packages/trader/src/AppV2/Layout/index.ts
@@ -1,3 +1,0 @@
-import Layout from './Layout';
-
-export default Layout;
