+++ packages/trader/src/AppV2/Layouts/index.ts
@@ -1,0 +1,3 @@
+import Layout from './Layouts';
+
+export default Layout;
