+++ packages/trader/src/index.tsx
@@ -3,7 +3,6 @@ import { getPositionsV2TabIndexFromURL, makeLazyLoader, moduleLoader, routes } f
 import { Loading } from '@deriv/components';
 import { TCoreStores } from '@deriv/stores/types';
 import { TWebSocket } from 'Types';
-import { useDtraderV2Flag } from '@deriv/hooks';
 
 type Apptypes = {
     passthrough: {
@@ -11,11 +10,6 @@ type Apptypes = {
         WS: TWebSocket;
     };
 };
-
-const AppLoader = makeLazyLoader(
-    () => moduleLoader(() => import(/* webpackChunkName: "trader-app", webpackPreload: true */ './App/index')),
-    () => <Loading />
-)() as React.ComponentType<Apptypes>;
 
 const AppV2Loader = makeLazyLoader(
     () => moduleLoader(() => import(/* webpackChunkName: "trader-app-v2", webpackPreload: true */ './AppV2/index')),
@@ -30,10 +24,7 @@ const AppV2Loader = makeLazyLoader(
 )() as React.ComponentType<Apptypes>;
 
 const App = ({ passthrough }: Apptypes) => {
-    const { dtrader_v2_enabled, load_dtrader_module } = useDtraderV2Flag();
-    if (load_dtrader_module) {
-        return dtrader_v2_enabled ? <AppV2Loader passthrough={passthrough} /> : <AppLoader passthrough={passthrough} />;
-    }
-    return <Loading />;
+    return <AppV2Loader passthrough={passthrough} />;
 };
+
 export default App;
